Korea's Trade Data

trade_data returns a dataset of Korea's export and import data in a country-year format. It includes detailed information about the trade activities, such as export and import values by country and year, along with total export and import figures.

Usage

trade_data

Format

A data frame with 12 variables:

iso3c

A character vector of ISO 3166-1 alpha-3 country codes.

country

A character vector of country names corresponding to the iso3c codes.

year

A numeric vector of years for which trade data is available.

export

A numeric vector representing the USD value of exports from Korea to the respective countries in each year. The default is the value reported by the Korean Statistical Information Service. It was imputed from the Correlates of War project data when the former was missing.

import

A numeric vector representing the USD value of imports to Korea from the respective countries in each year. The default is the value reported by the Korean Statistical Information Service.It was imputed from the Correlates of War project data when the former was missing.

total_export

A numeric vector indicating the total value of Korea's exports in each year as reported by Korean Statistical Information Service.

total_import

A numeric vector indicating the total value of Korea's imports in each year as reported by Korean Statistical Information Service.

export_kosis

A numeric vector representing export values from Korea to each country as recorded by the Korean Statistical Information Service.

import_kosis

A numeric vector representing import values to Korea from each country as recorded by the Korean Statistical Information Service.

export_cow

A numeric vector representing export values from Korea to each country as recorded by the Correlates of War project.

import_cow

A numeric vector representing import values to Korea from each country as recorded by the Correlates of War project.

export_cons_2015

A numeric vector representing the 2015 constant USD value of exports from Korea to the respective countries in each year. The figures are inflation-adjusted values of the export variable. Inflation adjustment is done using US Census Bureau's R-CPI-U-RS price series.

import_cons_2015

A numeric vector representing the 2015 constant USD value of imports to Korea from the respective countries in each year. The figures are inflation-adjusted values of the import variable. Inflation adjustment is done using US Census Bureau's R-CPI-U-RS price series.

total_export_cons_2015

A numeric vector representing the 2015 constant USD value of Korea's total exports in each year. The figures are inflation-adjusted values of the total_export variable. Inflation adjustment is done using US Census Bureau's R-CPI-U-RS price series.

total_import_cons_2015

A numeric vector representing the 2015 constant USD value of Korea's total imports in each year. The figures are inflation-adjusted values of the total_import variable. Inflation adjustment is done using US Census Bureau's R-CPI-U-RS price series.

updated_at

A date vector indicating the last update date of the data.

Source

https://kosis.kr/statHtml/statHtml.do?orgId=360&tblId=DT_1R11006_FRM101&conn_path=I3

https://correlatesofwar.org/data-sets/bilateral-trade/

Value

Returns a dataframe of Korea's export and import in a country-year format.